+++ tests/actions/conftest.py
@@ -2,8 +2,8 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from secureli.services.language_analyzer import AnalyzeResult
 from secureli.services.language_support import LanguageMetadata
+from secureli.shared.models.language import AnalyzeResult
 
 
 # Register generic mocks you'd like available for every test.
