+++ tests/actions/test_initializer_action.py
@@ -5,11 +5,9 @@ import pytest
 
 from secureli.actions.action import ActionDependencies
 from secureli.actions.initializer import InitializerAction
-from secureli.repositories.secureli_config import SecureliConfig
-from secureli.repositories.settings import SecureliFile, TelemetrySettings
-from secureli.services.language_config import LanguageNotSupportedError
 from secureli.services.logging import LogAction
 from secureli.settings import Settings
+from secureli.shared.models.language import LanguageNotSupportedError
 
 test_folder_path = Path("does-not-matter")
